cv_service/requirements.txt
flask 
mediapipe==0.10.14 
opencv-python-headless 
numpy 
scikit-learn 
joblib
flask-cors 
pillow
gunicorn==23.0.0
python-dotenv